MM-11697: edits per PR feedback
